+++ docker_tests/test_prod_image.py
@@ -85,7 +85,6 @@ class TestPythonPackages:
             lines = PREINSTALLED_PROVIDERS
         else:
             lines = (d.strip() for d in INSTALLED_PROVIDER_PATH.read_text().splitlines())
-            lines = (d for d in lines)
         packages_to_install = {f"apache-airflow-providers-{d.replace('.', '-')}" for d in lines}
         assert len(packages_to_install) != 0